+++ eslint.config.ts
@@ -150,7 +150,7 @@ export default eslintConfig({
             patterns: [
               {
                 regex:
-                  '^(?:local-pkg|type-fest|klona|@antfu/utils|defu|es-toolkit|remeda|node:util)$',
+                  '^(?:local-pkg|type-fest|klona|@antfu/utils|defu|es-toolkit|remeda|node:util|destr|string-ts)$',
                 message:
                   'Please do not use this package directly, import utility functions from `utils.ts` instead',
               },