+++ src/config.ts
@@ -769,6 +769,7 @@ export const eslintConfigInternal = async (
 
   const usedPluginPrefixes: readonly PluginPrefix[] = loadPluginsOnDemand
     ? // Sorting ensures that plugins will be present in the resulting config in the consistent order every time
+      // eslint-disable-next-line unicorn/no-array-sort
       [...context.usedPlugins].sort()
     : LOADABLE_PLUGIN_PREFIXES_LIST;
   const usedParserPrefixes = [...context.usedParsers.keys()];